+++ packages/client-auth/src/sdk/clients.ts
@@ -95,6 +95,7 @@ export class Clients extends ClientSDK {
                 headers: headers$,
                 query: query$,
                 body: body$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );
@@ -129,13 +130,13 @@ export class Clients extends ClientSDK {
             .match(response, { extraFields: responseFields$ });
 
         const nextFunc = (responseData: unknown): Paginator<operations.ClientsListResponse> => {
-            const nextCursor = dlv(responseData, "cursornext");
+            const nextCursor = dlv(responseData, "cursor.next");
 
             if (nextCursor == null) {
                 return () => null;
             }
 
-            const results = dlv(responseData, "cursordata");
+            const results = dlv(responseData, "cursor.data");
             if (!results.length) {
                 return () => null;
             }
@@ -190,6 +191,7 @@ export class Clients extends ClientSDK {
                 headers: headers$,
                 query: query$,
                 body: body$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );
@@ -269,6 +271,7 @@ export class Clients extends ClientSDK {
                 headers: headers$,
                 query: query$,
                 body: body$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );
@@ -351,6 +354,7 @@ export class Clients extends ClientSDK {
                 headers: headers$,
                 query: query$,
                 body: body$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );
@@ -430,6 +434,7 @@ export class Clients extends ClientSDK {
                 headers: headers$,
                 query: query$,
                 body: body$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );