+++ packages/client-auth/src/sdk/sdk.ts
@@ -85,6 +85,7 @@ export class AuthClient extends ClientSDK {
                 path: path$,
                 headers: headers$,
                 query: query$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );