+++ packages/client-auth/src/sdk/users.ts
@@ -92,6 +92,7 @@ export class Users extends ClientSDK {
                 headers: headers$,
                 query: query$,
                 body: body$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );
@@ -126,13 +127,13 @@ export class Users extends ClientSDK {
             .match(response, { extraFields: responseFields$ });
 
         const nextFunc = (responseData: unknown): Paginator<operations.UsersListResponse> => {
-            const nextCursor = dlv(responseData, "cursornext");
+            const nextCursor = dlv(responseData, "cursor.next");
 
             if (nextCursor == null) {
                 return () => null;
             }
 
-            const results = dlv(responseData, "cursordata");
+            const results = dlv(responseData, "cursor.data");
             if (!results.length) {
                 return () => null;
             }
@@ -194,6 +195,7 @@ export class Users extends ClientSDK {
                 headers: headers$,
                 query: query$,
                 body: body$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );